[INLONG-10096][Manager] Fix error
